Add synchonization for storage_sharing file tracker
It is required because multiple threads can access to the tracker,
even though we assume that there's only one thread for each file
at one time.

If we don't assume that there's only one thread can access for same
file (not tracker), we need to add more synchronization logic in
file class.