Implement Shadow DDES
Bootstrap master is implemented, which works as a tracker
among Shadow systems.
Currently the host is assigned fairly among the systems.
The Shadow systems are connected as a full mesh network,
sending packet event for corresponding host-assigned system.
At the end of each round, the consensus for the time window
begins among the systems.